requirements.txt
langchain
langchain-community
langchain-core
openai
pypdf
tiktoken
aiofiles
PyPDF2
chromadb
python-dotenv
streamlit
-e .